requirements.txt
tensorflow==2.20.0
streamlit>=1.31.0
numpy>=1.26.0
pillow>=10.0.0
opencv-python-headless>=4.9.0